requirements.txt
Django==5.1.4
djangorestframework==3.16.1
django-cors-headers==4.4.0
